split view code out into separate .js file
