+++ js/adapt-scrollPrompt.js
@@ -1,48 +1,7 @@
-define(function(require) {
-
-	var Adapt = require('coreJS/adapt');
-	var Backbone = require('backbone');
-
-	var scrollPromptView = Backbone.View.extend({
-
-		className: 'scrollPrompt',
-
-		events: {
-			"click .js-scrollPrompt-btn": "onScrollPromptClick"
-		},
-
-		initialize: function() {
-		var scrollPrompt = this.model.get('_scrollPrompt');
-
-		if (!scrollPrompt || !scrollPrompt._isEnabled) return;
-			this.render();
-		},
-
-		render: function() {
-			var data = this.model.toJSON();
-			var template = Handlebars.templates['scrollPrompt'];
-
-			this.$el.html(template(data))
-			_.defer(_.bind(this.postRender, this));
-		},
-
-		postRender: function() {
-			this.listenTo(Adapt, 'remove', this.remove);
-		},
-
-		onScrollPromptClick: function(event) {
-			var type = this.model.get("_type");
-			/* set scroll to selector depending on model type */
-			if (type === 'course') {
-				Adapt.scrollTo('.js-children', { duration: 800 });
-			} else if (type === 'page'){
-				Adapt.scrollTo('.article', { duration: 800 });
-			}
-		}
-
-	});
-
-	Adapt.on('app:dataReady', function() {
+define([
+  'core/js/adapt',
+  './scrollPromptView'
+], function(Adapt, ScrollPromptView) {
 
 		Adapt.on('menuView:ready pageView:ready', function(view) {
 
@@ -51,23 +10,16 @@ define(function(require) {
 			var scrollPrompt = model.get('_scrollPrompt');
 			if (!scrollPrompt || !scrollPrompt._isEnabled) return;
 
-			var modelType;
+    var modelType = model.get('_type');
+    if (modelType === 'course') modelType = 'menu';
 
-			switch (model.get("_type")) {
-				case "page":
-					modelType = "page";
-					break;
-				case "course":
-					modelType = "menu";
-			}
-			/* set model type selector to append scroll prompt */
+    /* set model type selector to append scroll prompt to */
 			var modelTypeSelector = '.' + modelType + '__header-inner';
 
-			new scrollPromptView({
-				model: view.model
+    new ScrollPromptView({
+      model: model
 			}).$el.appendTo(view.$(modelTypeSelector));
 
 		});
-	});
 
 });
